+++ cdap-app-fabric/src/main/java/co/cask/cdap/app/guice/DistributedProgramRunnableModule.java
@@ -39,16 +39,12 @@ import co.cask.cdap.data2.registry.RuntimeUsageRegistry;
 import co.cask.cdap.explore.guice.ExploreClientModule;
 import co.cask.cdap.internal.app.queue.QueueReaderFactory;
 import co.cask.cdap.internal.app.store.remote.RemoteLineageWriter;
-import co.cask.cdap.internal.app.store.remote.RemotePrivilegesFetcher;
 import co.cask.cdap.internal.app.store.remote.RemoteRuntimeStore;
 import co.cask.cdap.internal.app.store.remote.RemoteRuntimeUsageRegistry;
 import co.cask.cdap.logging.guice.LoggingModules;
 import co.cask.cdap.metrics.guice.MetricsClientRuntimeModule;
 import co.cask.cdap.notifications.feeds.client.NotificationFeedClientModule;
-import co.cask.cdap.security.authorization.AuthorizationEnforcementService;
-import co.cask.cdap.security.authorization.DefaultAuthorizationEnforcementService;
-import co.cask.cdap.security.spi.authorization.AuthorizationEnforcer;
-import co.cask.cdap.security.spi.authorization.PrivilegesFetcher;
+import co.cask.cdap.security.authorization.AuthorizationEnforcementModule;
 import com.google.inject.AbstractModule;
 import com.google.inject.Module;
 import com.google.inject.PrivateModule;
@@ -97,6 +93,7 @@ public class DistributedProgramRunnableModule {
       new AuditModule().getDistributedModules(),
       new NamespaceClientRuntimeModule().getDistributedModules(),
       new AuthorizationModule(),
+      new AuthorizationEnforcementModule().getDistributedModules(),
       new AbstractModule() {
         @Override
         protected void configure() {
@@ -110,14 +107,6 @@ public class DistributedProgramRunnableModule {
 
           // For binding StreamWriter
           install(createStreamFactoryModule());
-
-          // also bind AuthorizationEnforcementService as a singleton. This binding is used while starting/stopping
-          // the service itself.
-          bind(AuthorizationEnforcementService.class).to(DefaultAuthorizationEnforcementService.class)
-            .in(Scopes.SINGLETON);
-          // bind AuthorizationEnforcer to AuthorizationEnforcementService
-          bind(AuthorizationEnforcer.class).to(AuthorizationEnforcementService.class).in(Scopes.SINGLETON);
-          bind(PrivilegesFetcher.class).to(RemotePrivilegesFetcher.class);
         }
       }
     );
