+++ cdap-app-fabric/src/main/java/co/cask/cdap/gateway/handlers/meta/RemoteLineageWriterHandler.java
@@ -16,9 +16,9 @@
 
 package co.cask.cdap.gateway.handlers.meta;
 
+import co.cask.cdap.common.internal.remote.MethodArgument;
 import co.cask.cdap.data2.metadata.lineage.AccessType;
 import co.cask.cdap.data2.metadata.writer.LineageWriter;
-import co.cask.cdap.internal.app.store.remote.MethodArgument;
 import co.cask.cdap.proto.Id;
 import co.cask.http.HttpResponder;
 import com.google.inject.Inject;
