+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/guice/AppFabricTestModule.java
@@ -40,7 +40,6 @@ import co.cask.cdap.explore.guice.ExploreClientModule;
 import co.cask.cdap.gateway.handlers.meta.RemoteSystemOperationsServiceModule;
 import co.cask.cdap.internal.app.runtime.schedule.Scheduler;
 import co.cask.cdap.internal.app.runtime.schedule.SchedulerException;
-import co.cask.cdap.internal.app.store.remote.RemotePrivilegesFetcher;
 import co.cask.cdap.logging.guice.LogReaderRuntimeModules;
 import co.cask.cdap.logging.guice.LoggingModules;
 import co.cask.cdap.metadata.MetadataServiceModule;
@@ -50,8 +49,8 @@ import co.cask.cdap.notifications.feeds.guice.NotificationFeedServiceRuntimeModu
 import co.cask.cdap.notifications.guice.NotificationServiceRuntimeModule;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.ScheduledRuntime;
+import co.cask.cdap.security.authorization.AuthorizationEnforcementModule;
 import co.cask.cdap.security.guice.SecureStoreModules;
-import co.cask.cdap.security.spi.authorization.PrivilegesFetcher;
 import co.cask.cdap.store.guice.NamespaceStoreModule;
 import com.google.common.collect.ImmutableList;
 import com.google.inject.AbstractModule;
@@ -118,13 +117,9 @@ public final class AppFabricTestModule extends AbstractModule {
     install(new MetadataServiceModule());
     install(new RemoteSystemOperationsServiceModule());
     install(new AuthorizationModule());
+    // we want to use RemotePrivilegesFetcher in this module, since app fabric service is started
+    install(new AuthorizationEnforcementModule().getStandaloneModules());
     install(new SecureStoreModules().getInMemoryModules());
-    install(new AbstractModule() {
-      @Override
-      protected void configure() {
-        bind(PrivilegesFetcher.class).to(RemotePrivilegesFetcher.class);
-      }
-    });
   }
 
   private Scheduler createNoopScheduler() {
